#56 merge fixes
