+++ frontend/package.json
@@ -26,7 +26,7 @@
     "@angular/compiler": "^21.2.10",
     "@angular/core": "^21.2.10",
     "@angular/forms": "^21.2.10",
-    "@angular/material": "^21.2.10",
+    "@angular/material": "^21.2.8",
     "@angular/platform-browser": "^21.2.10",
     "@angular/platform-server": "^21.2.10",
     "@angular/router": "^21.2.10",
