--- frontend/src/app/components/curriculums/curriculum-planning/curriculum-planning.css
@@ -1,160 +1,0 @@
-.planning {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-  min-height: 100%;
-}
-
-.planning__header {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 16px;
-}
-
-.planning__heading {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.planning__title {
-  font-family: 'Lexend', system-ui, sans-serif;
-  font-size: 28px;
-  font-weight: 700;
-  line-height: 1;
-  letter-spacing: 0;
-  margin: 0;
-  text-transform: uppercase;
-  color: var(--color-page-title);
-}
-
-.planning__save {
-  background: var(--color-button);
-  color: var(--color-button-text);
-  border: none;
-  border-radius: 8px;
-  padding: 10px 20px;
-  font-size: 14px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 120ms ease;
-}
-
-.planning__save:hover:not(:disabled) {
-  background: var(--color-button-hover);
-}
-
-.planning__save:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
-
-.planning__panels {
-  display: grid;
-  grid-template-columns: minmax(0, 1fr) 1px minmax(0, 1fr);
-  gap: 32px;
-  flex: 1;
-  align-items: stretch;
-  border-left: 1px solid var(--color-divider);
-  padding-left: 32px;
-}
-
-.planning__panel {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  min-height: 400px;
-}
-
-.planning__section-title {
-  font-size: 13px;
-  font-weight: 700;
-  margin: 0;
-  text-transform: uppercase;
-  letter-spacing: 0.06em;
-  color: var(--color-section-title);
-}
-
-.planning__divider {
-  width: 1px;
-  background: var(--color-divider);
-  align-self: stretch;
-}
-
-th {
-  color: var(--color-light-text);
-  font-size: 14px;
-  font-weight: 200;
-}
-
-td {
-  font-size: 12px;
-  font-weight: 600;
-}
-
-.course__table {
-  border-collapse: separate;
-  border-spacing: 0 6px;
-  table-layout: fixed;
-}
-
-.course__row td {
-  background-color: var(--color-table-background);
-}
-
-.course__row td:first-child {
-  border-radius: 8px 0 0 8px;
-}
-
-.course__row td:nth-last-child(2) {
-  border-radius: 0 8px 8px 0;
-}
-
-.col__title {
-  padding-left: 8px;
-  width: 65%;
-  text-align: left;
-}
-
-.col__credits {
-  width: 15%;
-  text-align: center;
-}
-
-.col__semester {
-  width: 15%;
-  text-align: center;
-}
-
-.badge__center {
-  display: flex;
-  justify-content: center;
-}
-
-.col__action {
-  width: 5%;
-  text-align: left;
-  padding-left: 8px;
-}
-
-.add__btn {
-  background: none;
-  border: none;
-  font-size: 20px;
-  cursor: pointer;
-  padding: 8px;
-}
-
-.add__btn:hover {
-  background: var(--color-table-background);
-  border-radius: 8px;
-}
-
-.course__row td:last-child {
-  background: none;
-}
-
-.course__code {
-  font-weight: 400;
-}
