--- frontend/src/app/components/curriculums/curriculum-planning/curriculum-planning.html
@@ -1,63 +1,0 @@
-<section class="planning">
-  <header class="planning__header">
-    <h1 class="planning__title">{{ 'CurriculumPlanning.Title' | t }}</h1>
-    <button type="button" class="planning__save" disabled>
-      {{ 'Common.Save' | t }}
-    </button>
-  </header>
-
-  <div class="planning__panels">
-    <section class="planning__panel">
-      <h2 class="planning__section-title">
-        {{ 'CurriculumPlanning.SemesterPlan' | t }}
-      </h2>
-    </section>
-
-    <div class="planning__divider" aria-hidden="true"></div>
-
-    <section class="planning__panel">
-      <h2 class="planning__section-title">
-        {{ 'CurriculumPlanning.Subjects' | t }}
-      </h2>
-      <app-search-bar
-        [placeholder]="'CurriculumPlanning.SearchPlaceholder' | t"
-        (searchChange)="searchTitle.set($event)"
-      >
-      </app-search-bar>
-      @if (courses().length === 0) {
-        <app-notification [message]="'CurriculumPlanning.NoCoursesNotification' | t" />
-      } @else {
-        <table class="course__table">
-          <thead>
-            <tr>
-              <th class="col__title"></th>
-              <th class="col__credits">{{ 'CurriculumPlanning.CoursePoints' | t }}</th>
-              <th class="col__semester">{{ 'CurriculumPlanning.StudySemester' | t }}</th>
-              <th class="col__action"></th>
-            </tr>
-          </thead>
-          <tbody>
-            @for (course of courses(); track course.oisExternalId) {
-              <tr class="course__row">
-                <td class="col__title">
-                  {{ course.titleEt }} <span class="course__code">({{ course.code }})</span>
-                </td>
-                <td class="col__credits">{{ course.credits }}</td>
-                <td class="col__semester">
-                  @if (course.semesterType) {
-                    <div class="badge__center">
-                      <app-semester-badge [semester]="course.semesterType" />
-                    </div>
-                  }
-                </td>
-                <td class="col__action">
-                  <button type="button" class="add__btn">+</button>
-                </td>
-              </tr>
-            }
-          </tbody>
-        </table>
-      }
-    </section>
-  </div>
-</section>
