--- frontend/src/app/components/curriculums/curriculums.css
@@ -1,21 +1,0 @@
-.curriculums__header {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 16px;
-}
-
-.curriculums__add {
-  background: var(--color-button);
-  color: var(--color-button-text);
-  text-decoration: none;
-  border-radius: 8px;
-  padding: 10px 20px;
-  font-size: 14px;
-  font-weight: 600;
-  transition: background-color 120ms ease;
-}
-
-.curriculums__add:hover {
-  background: var(--color-button-hover);
-}
