--- frontend/src/app/components/curriculums/curriculums.html
@@ -1,4 +1,0 @@
-<header class="curriculums__header">
-  <app-placeholder-page titleKey="Curriculums.Title" />
-  <a routerLink="new" class="curriculums__add">{{ 'Common.AddNew' | t }}</a>
-</header>
