--- frontend/src/app/components/dashboard/dashboard.html
@@ -1,1 +1,0 @@
-<app-placeholder-page titleKey="Dashboard.Title" />
