--- frontend/src/app/components/dashboard/dashboard.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { PlaceholderPage } from '@/components';
-
-@Component({
-  selector: 'app-dashboard',
-  imports: [PlaceholderPage],
-  templateUrl: './dashboard.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class Dashboard {}
