--- frontend/src/app/components/schedule/schedule.html
@@ -1,1 +1,0 @@
-<app-placeholder-page titleKey="Schedule.Title" />
