--- frontend/src/app/components/schedule/schedule.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { PlaceholderPage } from '@/components';
-
-@Component({
-  selector: 'app-schedule',
-  imports: [PlaceholderPage],
-  templateUrl: './schedule.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class Schedule {}
