--- frontend/src/app/components/settings/settings.html
@@ -1,1 +1,0 @@
-<app-placeholder-page titleKey="Settings.Title" />
