--- frontend/src/app/components/studies/studies.html
@@ -1,1 +1,0 @@
-<app-placeholder-page titleKey="Studies.Title" />
