--- frontend/src/app/components/studies/studies.ts
@@ -1,10 +1,0 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { PlaceholderPage } from '@/components';
-
-@Component({
-  selector: 'app-studies',
-  imports: [PlaceholderPage],
-  templateUrl: './studies.html',
-  changeDetection: ChangeDetectionStrategy.OnPush,
-})
-export class Studies {}
