+++ frontend/src/app/features/curriculums/curriculum-planning/curriculum-planning.html
@@ -19,6 +19,45 @@
       <h2 class="planning__section-title">
         {{ 'CurriculumPlanning.Subjects' | t }}
       </h2>
+      <app-search-bar
+        [placeholder]="'CurriculumPlanning.SearchPlaceholder' | t"
+        (searchChange)="searchTitle.set($event)"
+      >
+      </app-search-bar>
+      @if (courses().length === 0) {
+        <app-notification [message]="'CurriculumPlanning.NoCoursesNotification' | t" />
+      } @else {
+        <table class="course__table">
+          <thead>
+            <tr>
+              <th class="col__title"></th>
+              <th class="col__credits">{{ 'CurriculumPlanning.CoursePoints' | t }}</th>
+              <th class="col__semester">{{ 'CurriculumPlanning.StudySemester' | t }}</th>
+              <th class="col__action"></th>
+            </tr>
+          </thead>
+          <tbody>
+            @for (course of courses(); track course.oisExternalId) {
+              <tr class="course__row">
+                <td class="col__title">
+                  {{ course.titleEt }} <span class="course__code">({{ course.code }})</span>
+                </td>
+                <td class="col__credits">{{ course.credits }}</td>
+                <td class="col__semester">
+                  @if (course.semesterType) {
+                    <div class="badge__center">
+                      <app-semester-badge [semester]="course.semesterType" />
+                    </div>
+                  }
+                </td>
+                <td class="col__action">
+                  <button type="button" class="add__btn">+</button>
+                </td>
+              </tr>
+            }
+          </tbody>
+        </table>
+      }
     </section>
   </div>
 </section>
