--- frontend/src/app/pipes/t.pipe.ts
@@ -1,11 +1,0 @@
-import { Pipe, PipeTransform, inject } from '@angular/core';
-import { I18nService } from '../services/i18n.service';
-
-@Pipe({ name: 't', pure: false })
-export class TPipe implements PipeTransform {
-  private readonly i18n = inject(I18nService);
-
-  transform(key: string): string {
-    return this.i18n.translate(key);
-  }
-}
