+++ frontend/src/app/shared/components/index.ts
@@ -1,5 +1,5 @@
-export { Notification } from './Notification/notification';
-export { SemesterBadge } from './SemesterBadge/semester-badge';
-export { Search } from './Search/search';
+export { Notification } from './notification/notification';
+export { SemesterBadge } from './semester-badge/semester-badge';
+export { Search } from './search/search';
 export { PlaceholderPage } from './placeholder-page/placeholder-page';
 export { Layout } from './layout/layout';
