+++ frontend/src/app/shared/components/notification/notification.css
