+++ frontend/src/app/shared/components/search/search.css
