+++ frontend/src/app/shared/components/semester-badge/semester-badge.html
