+++ frontend/src/app/shared/components/semester-badge/semester-badge.spec.ts
