+++ frontend/src/app/shared/components/semester-badge/semester-badge.ts
