+++ frontend/src/app/shared/pipes/index.ts
